requirements.txt
streamlit
opencv-python
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
joblib==1.3.2
scikit-learn>=1.0.0
tensorflow==2.15.0
keras==2.15.0